Changed internal location of Dynabench iterators (no change to import API)
