+++ src/dynabench/dataset/__init__.py
@@ -3,7 +3,6 @@ Module for loading the data.
 """
 
 from ._download import download_equation
-from ._iterator import DynabenchIterator
-from ._simulation_iterator import DynabenchSimulationIterator
+from ._dynabench import DynabenchIterator, DynabenchSimulationIterator
 
 __all__ = ["download_equation", "DynabenchIterator", "DynabenchSimulationIterator"]
